Cleanup
- move re-used hooks to own file
- move re-used types and type guards to own files
- consistent button and th text
