+++ spa/INA-tool/src/components/DownloadConnectionButton.tsx
@@ -24,7 +24,7 @@ export function DownloadConnectionButton() {
   return (
     <Button variant="outline" onClick={downloadCSV}>
       <DownloadIcon />
-      Download as CSV
+      Download
     </Button>
   );
 }
