+++ spa/INA-tool/src/components/edges.tsx
@@ -1,60 +1,16 @@
 import {
+  type InnerStatementEdge,
+  type ActorDrivenConnection,
+  type OutcomeDrivenConnection,
+  type SanctionDrivenConnection,
+  drivenColors,
+} from "@/lib/edge";
+import {
   BaseEdge,
-  Edge,
   EdgeProps,
   getBezierPath,
   getStraightPath,
-  MarkerType,
 } from "@xyflow/react";
-
-/*
-Connections:
-Purple (Actor-driven connection):
-- (Animate) Direct/Indirect Object -> Attribute
-- Execution Constraint -> Attribute
-
-Green (Outcome-driven connection):
-- (Inanimate) Direct/Indirect Object -> Activation Condition
-
-Red (Sanction-driven connection):
-- Aim -> Activation Condition
-*/
-
-export type InnerStatementEdge = Edge<{ label?: string }, "inner-statement">;
-export type ActorDrivenConnection = Edge<
-  Record<string, unknown>,
-  "actor-driven"
->; // Purple
-export type OutcomeDrivenConnection = Edge<
-  Record<string, unknown>,
-  "outcome-driven"
->; // Green
-export type SanctionDrivenConnection = Edge<
-  Record<string, unknown>,
-  "sanction-driven"
->; // Red
-
-export type DrivenConnection = (
-  | ActorDrivenConnection
-  | OutcomeDrivenConnection
-  | SanctionDrivenConnection
-) & { uncompactSource?: string; uncompactTarget?: string };
-
-export type INAEdge = InnerStatementEdge | DrivenConnection;
-
-// eslint-disable-next-line react-refresh/only-export-components
-export function isInnerStatementEdge(
-  edge: INAEdge,
-): edge is InnerStatementEdge {
-  return edge.type === "inner-statement";
-}
-
-// eslint-disable-next-line react-refresh/only-export-components
-export function isDrivenConnectionEdge(
-  edge: INAEdge,
-): edge is DrivenConnection {
-  return !isInnerStatementEdge(edge);
-}
 
 export function InnerStatementEdge({
   id,
@@ -81,14 +37,6 @@ export function InnerStatementEdge({
     />
   );
 }
-
-// eslint-disable-next-line react-refresh/only-export-components
-export const drivenColors = {
-  "inner-statement": "#000000",
-  "actor-driven": "#a855f7",
-  "outcome-driven": "#22c55e",
-  "sanction-driven": "#ef4444",
-} as const;
 
 const actorDrivenStyle = {
   stroke: drivenColors["actor-driven"],
@@ -187,29 +135,3 @@ export const edgeTypes = {
   "outcome-driven": OutcomeDrivenConnection,
   "sanction-driven": SanctionDrivenConnection,
 } as const;
-
-function connectionMarkerEnd(type: keyof typeof edgeTypes) {
-  return {
-    type: MarkerType.Arrow,
-    width: 10,
-    height: 10,
-    color: drivenColors[type],
-  };
-}
-
-// eslint-disable-next-line react-refresh/only-export-components
-export function buildEdge(
-  sourceId: string,
-  targetId: string,
-  type: Exclude<INAEdge["type"], undefined>,
-): INAEdge {
-  return {
-    id: `${sourceId}-2-${targetId}`,
-    source: sourceId,
-    target: targetId,
-    type,
-    markerEnd: connectionMarkerEnd(type),
-    sourceHandle: type,
-    targetHandle: type,
-  };
-}
